config/src/test/java/org/n52/sensorweb/server/helgoland/adapters/config/ConfigurationReaderTest.java
/*
 * Copyright (C) 2015-2023 52°North Spatial Information Research GmbH
 *
 * This program is free software; you can redistribute it and/or modify it
 * under the terms of the GNU General Public License version 2 as published
 * by the Free Software Foundation.
 *
 * If the program is linked with libraries which are licensed under one of
 * the following licenses, the combination of the program with the linked
 * library is not considered a "derivative work" of the program:
 *
 *     - Apache License, version 2.0
 *     - Apache Software License, version 1.0
 *     - GNU Lesser General Public License, version 3
 *     - Mozilla Public License, versions 1.0, 1.1 and 2.0
 *     - Common Development and Distribution License (CDDL), version 1.0
 *
 * Therefore the distribution of the program linked with libraries licensed
 * under the aforementioned licenses, is permitted by the copyright holders
 * if the distribution is compliant with both the GNU General Public License
 * version 2 and the aforementioned licenses.
 *
 * This program is distributed in the hope that it will be useful, but
 * WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY
 * or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License
 * for more details.
 */
package org.n52.sensorweb.server.helgoland.adapters.config;

import javax.inject.Inject;

import org.junit.jupiter.api.Assertions;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.springframework.test.context.ContextConfiguration;
import org.springframework.test.context.junit.jupiter.SpringExtension;

@ExtendWith(SpringExtension.class)
@ContextConfiguration(classes = {ConfigurationReader.class})
public class ConfigurationReaderTest {

    @Inject
    private ConfigurationReader reader;

    @Test
    public void test_configuration() {
        Assertions.assertNotNull(reader.getDataSources());
        Assertions.assertEquals(4, reader.getDataSources().size());
        DataSourceConfiguration dsc = reader.getDataSources().get(0);
        Assertions.assertNotNull(dsc.getJobs());
        Assertions.assertEquals(2, dsc.getJobs().size());
        DataSourceConfiguration dsc2 = reader.getDataSources().get(1);
        Assertions.assertNotNull(dsc2.getJobs());
        Assertions.assertEquals(1, dsc2.getJobs().size());
    }
}
